Disable tool tag validation on modifier stat builder
That validation was meant to reduce the NBT size of items a bit by skipping storing stats from modifiers that are not applicable. Most notably due to modifiers like diamond.
But ultimately its just lead to a lot of bugs from forgetting which tag is required to "support" a stat. Still undecided on whether to entirely remove the stat supports hook or not.

Removes the most obvious symptom of #5309 as that issue was caused by tags not being loaded when the player inventory syncs client side. Will push a more proper fix later.