Prevent rebuilding tool stats if any of the required data is missing
This fully solves #5309 as the problematic registry was item tags.
Also adds some future proofing in case an item stack is ever built pre-data loading.
Worth noting that many of those getters do not consider the concept of their data being reloaded when switching words; its possible some data might be marked as loaded because it was loaded in a different world